CORE: added framework for additional effect proc weapons via scripting fixed a warning message every time rapid shot procced added additional effect procs for ranged WS corrected some instances of damage being uint16 instead of uint32, capped damage to 99999 fixed subeffects not displaying if submessage was 0 fixed adding/removing of samba daze effects fixed inter-party interactions between dancer sambas (samba only works for dnc in your party) fixed drain differences between drain/aspir samba i/ii/ii fixed functionality of haste samba moved enhancing sword damage bonus to modifier revised day/weather bonus for enspell damage removed hard coded additional effects in core (to be moved to scripts eventually) SCRIPTS: fixed some oddities in samba abilities scripts removed enspell mod from samba effects to interact correcetly with enspells added functions in magic.lua for additional effect calculations clamped damage to 99999 in magic.lua added additional effect scripts for bloody bolt, acid bolt, sleep bolt, sirocco kukri SQL: added additional effect mods to item_mods, added enspell damage bonus to enhancing sword corrected bonus to haste samba merits added custom status effect for haste samba haste bonus

+++ src/map/status_effect_container.cpp
@@ -930,7 +930,7 @@ CStatusEffect* CStatusEffectContainer::GetStatusEffect(EFFECT StatusID)
 *                                                                       *
 ************************************************************************/
 
-CStatusEffect* CStatusEffectContainer::GetStatusEffect(EFFECT StatusID, uint16 SubID)
+CStatusEffect* CStatusEffectContainer::GetStatusEffect(EFFECT StatusID, uint32 SubID)
 {
 	for (uint16 i = 0; i < m_StatusEffectList.size(); ++i)
 	{
